Refactored based on feedback
